Final commit. All features implemented. No known bugs.
Commit to be used for final demo on Friday Aug 16th, 2012

+++ src/edu/purdue/rockapp/RockArrayAdapter.java
@@ -58,7 +58,7 @@ public class RockArrayAdapter extends ArrayAdapter<String> {
 			Location.distanceBetween(currentLoc.getLatitudeE6()/1e6, currentLoc.getLongitudeE6()/1e6, 
 					rock.getLat()/1e6, rock.getLon()/1e6, results);
 			
-			distanceView.setText(Float.toString(results[0]) + " meters away");
+			distanceView.setText(Integer.toString((int)results[0]) + " meters away");
 		}
 		
 		return rowView;
